Reuse local repo artifacts in repo-build (#93)
* Reuse local repo artifacts in repo-build

When running the repo-build script on an existing repository, reuse the
previous artifacts from that repo if they are still valid (i.e. matching
versions).

Rename the --force flag in the repo-build script to --local to avoid
confusion since the flag does not force rebuilding the local artifacts,
but only forces to build them locally when applicable.

Add a new `clean` target to the Makefile.

Make packages build workdirs unique per package version, so that they
don’t conflict with existing previous versions, should the build
directory be kept.

* Fix spacing in Makefile USAGE

* Fix typo

* Fix wrong argument index in repo-build

* Promote reuse message to section level