Improve source archive extraction
Behavior change
---------------

Previously, the topmost directories were always stripped when extracting
source archives, which was confusing when the archive contained more
than one top-level directory (see
<https://github.com/toltec-dev/toltec/pull/74#discussion_r495615907>).

This PR changes that behavior to a more prudent one, which finds the
longest folder prefix in the archive (potentially empty), and removes it
when extracting the archive’s files. For example, when extracting an
archive containing:

    a/b/c/file1.txt
    a/b/c/file2.txt

the `a/b/c/` prefix will be removed. When extracting an archive
containing:

    a/file1.txt
    b/file2.txt

both the `a/` and `b/` folders will be kept.

Changes
-------

* Implement the new behavior
* Document it in docs/package.md#sources
* Update oxide to take advantage of this