Fix handling packages with multiple dependencies (#121)
Previously only the first element of the $depends (resp. $conflicts)
array was used to build the list of dependencies (resp. conflicting
packages) of a packages, due to an error in the build script. This
caused an error in the generated package for `reboot-guard`, where the
`python` dependency was not automatically installed. This PR fixes that.

It also changes the version of `reboot-guard` from `1.0.1~1` to
`1.0.1-1` to bring it in line with other package versions.