Add VNSee 0.3.0
Since GitHub source code archives do not include submodules, I had to
download the libvncserver submodule separately. This source needs to be
placed in a specific subfolder of the $srcdir instead of the root, so we
need to override the automatic archive extraction behavior in this case.

To enable that, the two following changes are also included in this PR:

* Add an optional `noextract` field to the recipes to allow disabling the
  automatic extraction of downloaded archives.
* Add a `prepare()` section for running tasks on the source files before
  the build Docker container is started.