Implement a strategy for renaming packages
* Create an `old` section for deprecated or transitional packages. This
section is not included in the package listing.
* Add a `replaces` recipe field, equivalent to the `Provides`,
`Conflicts` and `Replaces` Debian fields combined.
* Use these new features to add an upgrade path for renaming
`minesweeper` to `mines`.

To test:

* Make a clean install of Toltec stable.
* Install `minesweeper` (version 0.0.1-14).
* Change the branch to <https://toltec-dev.org/dev/feature/renaming> in
`/opt/etc/opkg.conf`.
* Run `opkg update` then `opkg upgrade`.

This seems to work on my device: the `mines` package is installed, then
the `minesweeper` package is removed. Opkg generates an error however,
which seems like it can be safely ignored.

```
Upgrading minesweeper on root from 0.0.1-14 to 0.0.1-15...
Downloading
https://toltec-dev.org/dev/feature/renaming/minesweeper_0.0.1-15_armv7-3.2.ipk
Installing mines (0.0.1-17) to root...
Downloading
https://toltec-dev.org/dev/feature/renaming/mines_0.0.1-17_armv7-3.2.ipk
Removing package minesweeper from root...
Configuring mines.
Collected errors:
 * pkg_run_script: Internal error: minesweeper has a NULL
 tmp_unpack_dir.
  * prerm_upgrade_old_pkg: prerm script for package "minesweeper" failed
```

I’m not sure if this is an Opkg bug or if I’m doing something wrong yet.