Fix missing prerm script in rm2fb
Due to a logic bug fixed in
<https://github.com/toltec-dev/toltec/pull/222/files#r559226589>,
prerm scripts were not generated for corresponding preremove functions.

rm2fb’s version was not increased since this fix was merged, therefore
the current package (1.0.0-6) does not contain a prerm script.

This fix increments rm2fb’s version to trigger a rebuild of the package.