Hardcode REMOTE_HTTP secret in PR workflows
The REMOTE_HTTP secret was introduced to make the remote server
configurable without changing the workflow code in #218. For security
reasons, PR workflows do not have access to secret variables (this is an
oversight of the aforementioned PR).

We could switch the PR workflow to trigger on the `pull_request_target`
event which allows access to secrets, but that would enable external
attackers to exfiltrate our secrets (see
<https://securitylab.github.com/research/github-actions-preventing-pwn-requests>).

The correct solution would be for GitHub to provide a mechanism for
marking secrets as non-sensitive (which is the case of REMOTE_HTTP). For
now, let’s just hardcode that secret in PR workflows.