Use xochitl instead of remarkable-shutdown in rm2fb-server
In the last software relase (2.6), new waveform modes were added to
xochitl to make the pinch-to-zoom feature work. Unfortunately, the
remarkable-shutdown binary was not updated to support those new
waveforms. As a consequence, using pinch-to-zoom with rm2fb results in a
less smooth experience than without rm2fb, since the screen gets
refreshed less frequently during the gesture.

See:

* <https://github.com/toltec-dev/toltec/issues/322#issuecomment-808235577>
* <https://github.com/ddvk/remarkable2-framebuffer/issues/56>

To address this issue, this PR changes the rm2fb-server to preload the
`/usr/bin/xochitl` binary instead of `/usr/bin/remarkable-shutdown` in
both the `rm2fb-server` command and `rm2fb.service`.

A start condition is added to `rm2fb.service` to only start if the
librm2fb_server library actually exists on the filesystem. This is to
avoid having two concurrent xochitl processes running, as the loader
ignores non-existent files provided in LD_PRELOAD.

Test plan:

* Manually run `rm2fb-server` and `rm2fb-client xochitl`, make use of
  the pinch-to-zoom gesture, and make sure no message of the form
  `Unable to complete update: invalid waveform ( 6 )` appears in the
  server process.
* Check that the display continuously updates when using the
  pinch-to-zoom gesture, even if the `rm2fb.service` service is running
  and xochitl is started with `rm2fb-client xochitl`.
* Remove the `/opt/lib/librm2fb_server.so.1` symlink and restart
  `rm2fb.service`. The service should not start. Make sure that there’s
  no duplicate `xochitl` process in this case.