Create GitHub release on push to stable (#335)
Resolves #290

A release will automatically be created for each push to the stable branch. The release ID will be in the ISO week format (e.g., `2021-W14-5`). The release title and message will be taken from the first commit of each push (the same that’s used for creating the Discord notification). A new `RELEASE_TOKEN` secret is added, containing a token with repo metadata read/write permission. This is needed because the default `GITHUB_TOKEN` does not have the permission to create releases.