Manage Opkg configuration with toltecctl
This is a three-headed monster of a PR, addressing three distinct yet linked issues.

New toltecctl features
----------------------

At its core, this PR extends the toltecctl script (provided by toltec-bootstrap, introduced in #356) and enables it to manage the Opkg configuration, i.e., switch branches and migrate the configuration when needed. To achieve this without risking to overwrite custom config directives added by users, a split configuration approach is taken (as suggested by @Eeems):

* The main `/opt/etc/opkg.conf` file is automatically generated by concatenating the files in the `/opt/etc/opkg.conf.d/` folder (in lexicographic order).
* The file `/opt/etc/opkg.conf.d/10-entware.conf` configures the Entware repository and architecture names.
* The file `/opt/etc/opkg.conf.d/15-toltec.conf` configures the Toltec repository and architecture names.
* Users may create new files in `/opt/etc/opkg.conf.d` to add their custom directives.

New toltecctl commands are made available to users for managing the Opkg configuration:

* `toltecctl generate-opkg-conf` - Regenerates the main config file and automatically runs `opkg update` if the config changed
* `toltecctl switch-branch [BRANCH]` - Switches the current Toltec branch, rebuilds the config file and runs `opkg update`

This new `toltecctl switch-branch` fixes #341. A confirmation prompt is shown when someone tries to switch to the `testing` branch to warn them of the potential instability.

Bootstrap script revamp
-----------------------

This fixes #141.

With the previously mentioned additions, toltecctl now almost fully duplicates the features of the bootstrap script: it can create bind mounts, initialize the Opkg configuration, and define the user’s PATH. This PR therefore also contains a new bootstrap script that fetches the `toltec-bootstrap` package (containing toltecctl) first, and then uses its functions to perform the install, instead of reimplementing the same logic. By consolidating the features and variables of the two scripts, I hope to reduce the surface for potential bugs. This also makes it easier to add migrations to the `toltec-bootstrap` package, since it shares the code of the bootstrap script that’s used for new installs.

Migration to the new Opkg configuration
---------------------------------------

Building on the toltecctl improvements, a config migration logic was added to the `toltec-bootstrap` package’s configure function. This will automatically migrate single-opkg.conf-file installs to the new split config layout, while keeping the currently configured branch.

This migration is also able to transition configs using the old repository layout (with a single folder containing all packages) to the new arch-subfolder layout. By publishing the new toltec-bootstrap package to the root folder of the stable branch, this will solve #370.