Hotfix: Fix local wget download issue in stable (#369)
The bootstrap script that was in stable created an empty
`/home/root/.local/bin/wget` file before downloading its contents using
`wget`. If a user happened to have `/home/root/.local/bin` in their PATH
before the `/usr/bin` folder, then the script would try to use the empty
`/home/root/.local/bin/wget` binary to download itself! (See #368.)

This issue is already fixed in testing, so this PR is just a backport of
this fix to the stable branch.