Fix toltecctl set-path command if .bashrc does not exist
The `toltecctl set-path` command is used to make sure that the user’s shell PATH is properly configured to see Toltec binaries. It first scans the existing .bashrc to remove old Toltec-related PATH definitions, and then appends its own definitions.

In the current implementation, if there’s no existing /home/root/.bashrc file, the script will exit with an error. This is a problem because `set-path` is called as part of the bootstrap script, so if someone runs bootstrap after doing a factory reset (which removes the .bashrc file), the bootstrap will fail.

This PR changes the implementation to skip cleaning the .bashrc if it doesn’t exist. It also fixes two wrong messages in the bootstrap script that refer to the `toltecctl enable` command, that is actually called `toltecctl reenable`.

Test plan: Remove (or rename) your .bashrc and upgrade toltec-bootstrap to the new version. Since `set-path` is called automatically when the package is upgraded, this should recreate a new .bashrc containing Toltec-related PATH definitions.

I tested this on my rM2.