Fix toltecctl set-path command if .bashrc does not exist (#372)
The `toltecctl set-path` command is used to make sure that the user’s shell PATH is properly configured to see Toltec binaries. It first scans the existing .bashrc to remove old Toltec-related PATH definitions, and then appends its own definitions.

In the current implementation, if there’s no existing /home/root/.bashrc file, the script will exit with an error. This is a problem because `set-path` is called as part of the bootstrap script, so if someone runs bootstrap after doing a factory reset (which removes the .bashrc file), the bootstrap will fail.

This PR changes the implementation to skip cleaning the .bashrc if it doesn’t exist. It also fixes two wrong messages in the bootstrap script that refer to the `toltecctl enable` command, that is actually called `toltecctl reenable`.