Stable merge for week 29 of 2021
Important notices:

* When upgrading, users should run `opkg update && opkg upgrade` **twice** because of the changes introduced by this merge. (#310)
* Toltec **no longer supports** OS releases **earlier than 2.6**. You can check your current OS version in the settings panel. Please upgrade to 2.6 or newer after upgrading your Toltec install. (#366)
* To re-enable Toltec, **run `toltecctl reenable`** instead of the previous `entware-reenable` script that has been removed with this update.

Major change: The repository is now split into architecture subfolders. The `rmall` subfolder contains packages that can run both on reMarkable 1 and reMarkable 2 without modification. The `rm1` and `rm2` subfolders contain packages that are specific to each of the two devices. This change requires updating the Opkg configuration to point to the appropriate architectures. The migration is automated by placing an upgrade to the `toltec-bootstrap` package in the root folder of the repo that contains a migration script. (Note that this package is not generated automatically by our build system but will be added manually by the server administrator.)

New packages:

* bash-completion - 2.11-3 (#277, #375)
* display - 1.0.1-7 _(replaces the server part of the old rm2fb package)_ (#310)
* fuse - 1.0.0-2 (#331)
* gocryptfs - 2.0-beta2 (#318)
* innernet - 1.3.1-2 (#350, #358, #361)
* netsurf - 0.4.0-1 (#339)
* puzzles - 0.2.2-2 (#326)
* remarkable-stylus - 0.0.3-1 (#383)
* rm2fb-client - 1.0.1-7 _(replaces the client part of the old rm2fb package)_ (#310)
* tilem - 0.0.7-1 (#352)
* toltec-completion - 0.2.0-1 (#277)
* yaft - 0.0.4-3 (#261, #353)
* zoneinfo-utils - 2021a-2 (#182)

Updated packages:

* appmarkable - 0.0.0-11
* bufshot - 0.1.0-5
* calculator - 0.0.0-15
* chessmarkable - 0.7.1-2 (#362, #373)
* ddvk-hacks - 23.02-1 (#385)
* draft - 0.2.0-20
* evtest - 1.34-3
* fbink fbdepth fbink-doom - 1.23.1-4
* fingerterm - 1.3.5-14
* genie - 0.1.5-3
* harmony - 0.1.3-3
* iago - 0.1.0-4
* keywriter - 0.1.0-3
* koreader - 2021.06-1 (#330, #348, #357, #378)
* lamp - 0.1.0-4
* libdlib libdlib-dev - 19.21-2
* libvncserver libvncclient libvncserver-dev - 0.9.13-2
* mines - 0.1.2-4
* nao - 0.1.3-3
* oxide erode fret rot tarnish decay corrupt anxiety - 2.2-1 (#382)
* plato - 0.9.17-2 (#344)
* quickjs - 2020.11.08-2
* recrossable - 0.0.0-7
* remarkable-splash - 1.0-5
* remux - 0.1.9-3
* restream - 1.1-2
* retris - 0.6.3-3
* rmservewacominput - 0.3.0-2
* simple - 0.1.4-3
* templatectl - 0.1.3-3
* toltec-bootstrap - 0.1.0-1 _(includes toltecctl)_ (#379)
* vnsee - 0.4.0-2 (#355)
* wikipedia - 0.1.0-3
* wireguard - 1.0.20210219-2
* xochitl - 0.0.0-7 (#383, #382, #386)
* zshelf - 0.3.1-3

Website: The source files and scripts used for generating the website have been transferred to a separate repository (<https://github.com/toltec-dev/web>). A website rebuild is triggered automatically from this repository’s workflow when the stable branch is updated, so that the displayed checksum of the bootstrap script is always accurate.