netsurf: Improve rM2 compatibility (#395)
This PR adds a missing dependency of netsurf on the display package. This dependency ensures that rM2 users have the rm2fb server installed before they start netsurf. This diff also adds the patch_rm2fb flag to the recipe, so that netsurf can be started from the command line on rM2 without prefixing it with rm2fb-client.