Improve Opkg completion
This PR builds upon the work of @LinusCDE on Bash completions for Opkg
and adds some missing features.

* Add completion of flag names and values.
* Only suggest installed packages when completing `opkg remove`.
* And completion of other subcommands involving package names (depends,
  download, info, what.*, configure, files, status, upgrade, flag)
* Add completion of search and compare-versions operators.

This makes the completion function fairly complete as far as I can see.

A future area of work will be to cache the list of packages, currently
it is re-generated for each call to the completion function, which takes
some noticeable time.