fuse: Fix on-demand loading after install on rM1 (#405)
In the current state of the fuse package, the module will not get
autoloaded if you try to use it right after installing. You can see this
on rM1 by installing gocryptfs and trying to decrypt a directory
(without rebooting after installing the package).

This PR fixes that by forcing the creation of `/dev/fuse` that is used
for on-demand loading of the fuse kernel module. If you install this new
version of the package without having fuse installed before, the module
should get autoloaded without needing a restart of the device.