package: linux-5.4-imx: Initial commit
This is the initial commit of the 5.4 kernel. This provides additional
bug fixes and security benefits over the outdated 4.19 kernel that is
shipped on the device. This still includes all of the original drivers
and features.

This also includes changes from the rM2 kernel upstreaming process,
increasing the quality of the driver code.

This simplifies the install process specified here:
https://remarkablewiki.com/tech/updatingkernel

Signed-off-by: Alistair Francis <alistair@alistair23.me>