koreader: Add service unit (#441)
This PR adds a systemd service unit for starting KOReader at boot time (see #437). If KOReader fails to start, it falls back to Xochitl. It also updates the Xochitl service definition to alias to `launcher.service`; this is to prevent situations where both KOReader and Xochitl are enabled to start at boot time (see #438).