Make Xochitl work even if rm2fb doesn’t start (#455)
The goal of this PR is to make Xochitl start on reMarkable 2 even in cases where the rm2fb server does not start. This happens for example if one disables rm2fb.service or (in a future rm2fb update) if it’s ran on an unsupported system version. Making sure that Xochitl always starts can prevent soft-bricking for users who did not write down their SSH password and did not set up SSH key authentication.

The Xochitl wrapper script will now only set LD_PRELOAD if the rm2fb server is actually running (detected by checking the existence of `/dev/shm/swtfb.01`).

To implement this, I modified the script so that it sources files in `/opt/etc/xochitl.env.d` instead of only parsing them as key/value lists. This enables those files to have some logic (such as checking the existence of a file) before setting environment variables.

Co-authored-by: Nathaniel van Diepen <Eeems@users.noreply.github.com>