[rmkit] upgrade remux with rM1 support on 2.9 (#450)
this brings remux up to date with rmkit-dev/rmkit@8254893

major improvement is proper support for touch gestures in remux on rM1 after rebooting. the main issue was the display was resizing after remux started, so the touch gestures were using the wrong display size. holding the center button on rM1 would cause remux to restart (and fix this problem), so there is a workaround available.