Stable merge for week 42 of 2021 (#477)
* toltecctl: Honour dependencies when uninstalling (#456)

`toltecctl uninstall` will now try to uninstall packages in the reverse order of their dependencies, so that a package can use binaries it depends on during its removal scripts (see #448). Packages in a dependency cycle are removed in an unspecified order.

* [rmkit] upgrade remux with rM1 support on 2.9 (#450)

this brings remux up to date with rmkit-dev/rmkit@8254893

major improvement is proper support for touch gestures in remux on rM1 after rebooting. the main issue was the display was resizing after remux started, so the touch gestures were using the wrong display size. holding the center button on rM1 would cause remux to restart (and fix this problem), so there is a workaround available.

* Update KOReader to 2021.10 (#467)

Co-authored-by: raisjn <70462544+raisjn@users.noreply.github.com>

* Add cache for dependencies install (#473)

* Add cache for dependencies install

Co-authored-by: Mattéo Delabre <1370040+matteodelabre@users.noreply.github.com>
Co-authored-by: raisjn <70462544+raisjn@users.noreply.github.com>
Co-authored-by: Mattéo Delabre <spam@delab.re>