Move installation and usage instructions to the website (#498)
Installation and usage instructions are currently duplicated between the
README of this repo and our website. This PR removes those instructions
from the README and directs users to the website instead.

This change will force new users to make one more click to access the
install instructions (should they land first on our GitHub repo instead
of the website), but it will free us from having to maintain two copies
of the same thing and from having to manually keep the bootstrap
checksum up to date in the README.

In the long run, I’d like for the website to become the main source of
information for end-users, and for our repos to hold information for
developers/contributors. This change is a first step towards that goal.

Co-authored-by: Nathaniel van Diepen <Eeems@users.noreply.github.com>