display: Release 0.0.13 (#531)
This release of remarkable2-framebuffer complements PR #455 and brings a
fix for a potential bug in the client (see ddvk/remarkable2-framebuffer#85).

The rm2fb service will now only open the IPC channels after it has
confirmed that all the required offsets for the current system release
are available. This can take a few seconds; the service will also notify
the systemd daemon when it is ready to accept requests, so that
dependent services such as Xochitl can correctly wait for the server to
be available before starting.