package: linux-mainline: Initial commit (#488)
This package provides a mainline Linux kernel for rM2. This is based on
current Linux mainline with some extra reMarkable patches on top. This
doesn't yet have the same support as the vendor kernel, but does have
most of the functionality.

This simplifies the install process specified here:
https://remarkablewiki.com/tech/updatingkernel

This PR also adds a new “kernel” section for packages that provide
kernels or modules. Existing packages that fit this description are
moved to the new section.

For now, switching between kernels is to be done manually by the user.
A future PR (#569) will make this easier.

Signed-off-by: Alistair Francis <alistair@alistair23.me>
Co-authored-by: Mattéo Delabre <git.matteo@delab.re>